// src/app/interfaces/window-data.interface.ts
export interface WindowData {
	window_nameCode: string;
	imageUrl: string;
	content: string;
	nameCode: string;
	width_ft: number;
	width_in: number;
	height_ft: number;
	height_in: number;
	captured_at: string;
	latitude: string;
	longitude: string;
	image_uri: string;
	thumb_uri: string;
	// Add any other properties your window objects have
}